Update README.md
**SteelPanTutor Flutter App**  

SteelPanTutor Flutter App is a feature-rich, cross-platform application designed for steelpan enthusiasts, musicians, and learners to explore the vibrant world of steelpan music. The app offers a diverse range of instruments, each carefully crafted to emulate the authentic sound and feel of traditional steelpans.  

### **Included Instruments**  
SteelPanTutor features an extensive collection of steelpan instruments:  
- **Two sets of Tenor Pans**:  
  - **4ths and 5ths layout**  
  - **3rd and 4ths layout** (Low D to High C)  
- **One set of Double Tenor Pans**  
- **One set of Double Second Pans**  
- **One set of Double Guitar Pans**  
- **One Tenor Bass Pan**  
- **One Six Bass Pan**  

### **Key Features**  
- **MIDI Import and Playback**:  
  Import MIDI files and play them directly within the app, enabling users to practice or enjoy steelpan renditions of their favorite songs.  
- **Interactive Playback**:  
  Visual note playback to guide users through the melodies and rhythms, making learning engaging and effective.  
- **Customizable Settings**:  
  Save preferences for personalized playback and practice sessions.  
- **Advanced Visualization**:  
  Dynamic and colorful note visualizations to enhance the user experience and aid in learning.  
- **Light and Dark Modes**:  
  Seamless theme switching for comfortable use in any lighting condition.  
- **Cross-Platform Support**:  
  Designed with Flutter, the app runs smoothly on Android, iOS, and desktop platforms, ensuring a consistent experience everywhere.  

### **Who Is It For?**  
SteelPanTutor is perfect for steelpan enthusiasts of all levels—whether you're a beginner learning your first melody or a seasoned player exploring new compositions. The app is also an excellent tool for educators looking to provide a digital learning platform for their students.  

SteelPanTutor brings the joy of steelpan music to your fingertips, making it easier than ever to learn, play, and appreciate this beautiful instrument. Explore the versatility of steelpans, practice with precision, and immerse yourself in the vibrant world of steelpan music with SteelPanTutor!